BTEX-1369: Add extender device support in T2
Reason for change: Add extender device(Pods)support in T2
Test Procedure: T2 component should compile and run with device extender support
                and without affecting the existing functionalities
Risks: Medium

DELIA-55024,RDKB-40824,RDKB-40798: T2 fails to support Xconf-Fix

Reason for change: To fix T2 fails to support Xconf communication
Test Procedure: T2 component should communicate with Xconf
                and without affecting the existing functionalities
Risks: Medium

(cherry picked from commit 326542fc86b4cc04615be088067700d01b23e9e5)

+++ source/bulkdata/profile.c
@@ -378,7 +378,7 @@ static void* CollectAndReport(void* data)
     }
     clock_gettime(CLOCK_REALTIME, &endTime);
     getLapsedTime(&elapsedTime, &endTime, &startTime);
-    T2Info("Elapsed Time for : %s = %lu.%lu (Sec.NanoSec)\n", profile->name, elapsedTime.tv_sec, elapsedTime.tv_nsec);
+    T2Info("Elapsed Time for : %s = %lu.%lu (Sec.NanoSec)\n", profile->name, (unsigned long)elapsedTime.tv_sec, elapsedTime.tv_nsec);
     if(ret == T2ERROR_SUCCESS && jsonReport)
     {
         free(jsonReport);
